+++ Swifties/Models/WeeklyChallengeChartData.swift
@@ -1,0 +1,14 @@
+//
+//  ChartData.swift
+//  Swifties
+//
+//  Created by Imac  on 4/10/25.
+//
+
+import Foundation
+
+struct WeeklyChallengeChartData: Identifiable {
+    let id = UUID()
+    let label: String
+    let count: Int
+}
